+++ lib/management/index.ts
@@ -20,6 +20,11 @@ import withOutboundApplication from './outboundapplication';
 import withOutboundSCIM from './outboundscim';
 import withDescoper from './descoper';
 import withManagementKey from './managementKey';
+import withEngine from './engine';
+import withList from './list';
+import withJWTTemplate from './jwttemplate';
+import withScopeClaimMapping from './scopeclaimmapping';
+import withAnalytics from './analytics';
 import { FGAConfig } from './types';
 
 /** Constructs a higher level Management API that wraps the functions from code-js-sdk */
@@ -45,6 +50,11 @@ const withManagement = (client: HttpClient, fgaConfig?: FGAConfig) => ({
   fga: WithFGA(client, fgaConfig),
   descoper: withDescoper(client),
   managementKey: withManagementKey(client),
+  engine: withEngine(client),
+  list: withList(client),
+  jwtTemplate: withJWTTemplate(client),
+  scopeClaimMapping: withScopeClaimMapping(client),
+  analytics: withAnalytics(client),
 });
 
 export default withManagement;